pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mononoqe"
version = "0.1.2"
description = "A simple python library to mix Neural Network and Quantum computing"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "fire==0.7.0",
    "perceval-quandela==1.1.0",
    "lightning==2.5",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "tensorboard==2.18.0",
    "dataclasses-json",
    "timm",
    "pandas",
    "numpy==2.3.5",
    "scikit-learn==1.8.0",
    "seaborn==0.13.2",
]

[project.urls]
Repository = "https://github.com/QuantumCommons/mononoqe"

[tool.setuptools.packages.find]
include = ["mononoqe*"]
exclude = ["resources*", "tests*"]
